Revamp how ExternalProcess works
